# wazeroir: less allocations