… internal/engine/compiler/compiler.go
… internal/engine/compiler/compiler_conditional_save_test.go
… internal/engine/compiler/compiler_controlflow_test.go
… internal/engine/compiler/compiler_conversion_test.go
… internal/engine/compiler/compiler_memory_test.go
… internal/engine/compiler/compiler_numeric_test.go
… internal/engine/compiler/compiler_post1_0_test.go
… internal/engine/compiler/compiler_stack_test.go
… internal/engine/compiler/compiler_vec_test.go
… internal/engine/compiler/engine.go
… internal/engine/compiler/impl_amd64.go
… internal/engine/compiler/impl_amd64_test.go
… internal/engine/compiler/impl_arm64.go
… internal/engine/compiler/impl_arm64_test.go
… internal/engine/compiler/impl_vec_amd64.go
… internal/engine/compiler/impl_vec_amd64_test.go
… internal/engine/compiler/impl_vec_arm64.go
… internal/engine/compiler/impl_vec_arm64_test.go
… internal/wazeroir/compiler.go
… internal/wazeroir/compiler_test.go
… internal/wazeroir/operations.go